# api/openapi/pda-v1.yaml
openapi: 3.0.3
info:
  title: PDA Backend API
  version: 0.3.0
  description: PDA public transport envelope, headers, metadata, identity, device, warehouse, and workflow contract.
servers:
  - url: /api/pda/v1
paths:
...
  /commands/{commandId}:
    get:
      operationId: commandStatus
      security: [{bearerAuth: []}]
      parameters: [{name: commandId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Canonical durable command status}, '403': {description: Command access denied}, '404': {description: Command not found}}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...